Add a new key binding CTRL+R which resolve, load and set the TP
This commit adds a new command + key binding to set the current .tpd
file as the current targetplatform. IT wraps the .target generation, the
target resolving and reload in a single job, with no need for opening
the preference page or the .target editor.

This command is also available right-clicking on a .tpd file (if only
one is selected).

Fix #37