Fix whitespace and remove useless feature metadata in maven targets
All other parts of the generated target use an indentation of 2 spaces.
Use the same in the Maven part. Also remove the useless metadata from
the generated (optional) feature. There is no need to have those tags,
and the generated nonsense data is end-user visible.

Add an assertion to verify the absence of tab characters to avoid
similar problems in the future.